fix(deprize): keep server-only Privy auth out of the browser test bundle
eligibility.cy.ts imported selectSessionWallet from sessionWallet, which
also holds walletFromSession and so pulls in next-auth and
@privy-io/server-auth. Bundling that spec for the browser loaded
@privy-io/server-auth, which throws "cannot be used in a browser
environment" on import and failed the whole component-test shard.

selectSessionWallet is pure, so move it to its own module. The three API
routes only ever imported walletFromSession, so they are unaffected.
