+++ ui/cypress/integration/lib/deprize/eligibility.cy.ts
@@ -1,7 +1,7 @@
 import { evaluateEligibility } from '@/lib/deprize/eligibility'
 import { isRestrictedJurisdiction } from '@/lib/deprize/restrictedJurisdictions'
 import { parseOfacEthList } from '@/lib/deprize/sanctions'
-import { selectSessionWallet } from '@/lib/deprize/sessionWallet'
+import { selectSessionWallet } from '@/lib/deprize/selectSessionWallet'
 import {
   checkVpnOrProxy,
   hostingLooksLikeProxy,
