+++ ui/lib/deprize/sessionWallet.ts
@@ -1,9 +1,8 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
 import { getServerSession } from 'next-auth/next'
-import { getAddress } from 'viem'
 import { getPrivyUserData } from '@/lib/privy'
 import { authOptions } from '@/pages/api/auth/[...nextauth]'
-import { isHexAddress } from './eligibility'
+import { selectSessionWallet } from './selectSessionWallet'
 
 function accessTokenFromRequest(req: NextApiRequest): string | null {
   const authHeader = req.headers.authorization
@@ -11,25 +10,6 @@ function accessTokenFromRequest(req: NextApiRequest): string | null {
     const token = authHeader.slice(7).trim()
     if (token) return token
   }
-  return null
-}
-
-/** Pick a session wallet. A claimed address must belong to the session. */
-export function selectSessionWallet(
-  wallets: string[],
-  claimedWallet?: string | null
-): string | null {
-  const owned = wallets.filter((addr) => isHexAddress(addr))
-  if (owned.length === 0) return null
-
-  const claimed = claimedWallet?.trim() || ''
-  if (claimed) {
-    if (!isHexAddress(claimed)) return null
-    const match = owned.find((w) => w.toLowerCase() === claimed.toLowerCase())
-    return match ? getAddress(match) : null
-  }
-
-  if (owned.length === 1) return getAddress(owned[0])
   return null
 }
